requirements.txt
pyttsx3
speechrecognition
psutil
google-generativeai
python-dotenv
